+++ src/main/java/com/zylitics/api/model/NewBuildResponse.java
@@ -1,16 +1,8 @@
 package com.zylitics.api.model;
 
-import java.util.List;
-
-public class NewBuildResponse {
+public class NewBuildResponse extends AbstractBuildResponse {
   
   private int buildId;
-  
-  private String status;
-  
-  private String error;
-  
-  private List<TestDetail> testDetails;
   
   public int getBuildId() {
     return buildId;
@@ -18,33 +10,6 @@ public class NewBuildResponse {
   
   public NewBuildResponse setBuildId(int buildId) {
     this.buildId = buildId;
-    return this;
-  }
-  
-  public String getStatus() {
-    return status;
-  }
-  
-  public NewBuildResponse setStatus(String status) {
-    this.status = status;
-    return this;
-  }
-  
-  public String getError() {
-    return error;
-  }
-  
-  public NewBuildResponse setError(String error) {
-    this.error = error;
-    return this;
-  }
-  
-  public List<TestDetail> getTestDetails() {
-    return testDetails;
-  }
-  
-  public NewBuildResponse setTestDetails(List<TestDetail> testDetails) {
-    this.testDetails = testDetails;
     return this;
   }
 }